Add evm to starknet address mapping (#720)
Time spent on this PR: 0.3

## Pull request type

Please check the type of change your PR introduces:

- [x] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

The Kakarot contract cannot know whether a starknet address is deployed
or not.

Resolves #719 #474

## What is the new behavior?

A mapping evm_address -> starknet_address has been added:

- returns the starknet_address when it's deployed
- returns 0 otherwise

## Note

I didn't replace all the `compute_starknet_address`, and did actually
the minimal changes,
since it is still possible in some case (e.g. transfer) to send ETH to a
non deployed account.

However, now that we have introduced a mapping, we could move away from
this deterministic mapping
to allow any starknet contract to claim ownership of an EVM address.

This is let to further dev, probably in cairo1.