fix: signature extraction in decode legacy tx (#781)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days.
Did you spend 0.5 days on this PR or rather 2 days?  -->

Time spent on this PR: 0.1 day

## Pull request type

<!-- Please try to limit your pull request to one type,
submit multiple pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?
R and S part of the signature are assumed to be 32 bytes.

<!-- Please describe the current behavior that you are modifying,
or link to a relevant issue. -->

Resolves [573](https://github.com/kkrt-labs/ef-tests/issues/573)

## What is the new behavior?
R and S are correctly decoded by taking the length of the values.