Put balance in Account instead of State (#807)
Time spent on this PR: 0.2

## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

The `model.State` maintains a dict for accounts balances.
This is a mis conception due the the false belief that "we don't want
to fetch an account just to send ETH". But there is no way to send ETH
without making a CALL (apart in SELFDESTRUCT).

## What is the new behavior?

The balance belongs to the Account instead (as in the ef tests json
description for example).

Note: this is a preliminary work to run Kakarot in standalone (no
syscall, ie. no Starknet) mode.