dev: move skip file locally to repo (#847)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days.
Did you spend 0.5 days on this PR or rather 2 days?  -->

Time spent on this PR: 0.1 day

## Pull request type

<!-- Please try to limit your pull request to one type,
submit multiple pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [x] Other (please describe): Testing

## What is the current behavior?
Skip file for ef-tests are taken for the ef-tests repo.

<!-- Please describe the current behavior that you are modifying,
or link to a relevant issue. -->

Resolves https://github.com/kkrt-labs/ef-tests/issues/617

## What is the new behavior?
Skip file for kakarot v0 is moved to the local repo.