tests/improve AUTH tests
seperated the method of test__auth in test_system_operations.cairo into two methods. one which initializes evm with authority as set and other with authority unset taking a scaffolding pathway. made relevant changes to the python file. added a method in helpers.cairo to initialize an evm frame with authority set. updated readme to reflect correct command for testing units which was likely a typo.