fix unsound get_felt_bitlength  (#1199)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days.
Did you spend 0.5 days on this PR or rather 2 days?  -->

Time spent on this PR:
1h

## Pull request type

<!-- Please try to limit your pull request to one type,
submit multiple pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying,
or link to a relevant issue. -->

Resolves #<Issue number>

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Assumes u128 in get_felt_bitlength to get rid of bitwises
- add a range check on the prover given value. The prover may still find
an offset that gives the correct bit value starting from the beginning
of the dw array (if the correct value is present anywhere in the
bytecode or memory of the program), but the function would an offset
which would be wrong. bit_length must be constrained to be in [0,128]
- Optimize slightly lookup in pow2.
- Optimize slightly get_u256 bitlength logic.
- Improve function documentation

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg"
height="34" align="absmiddle"
alt="Reviewable"/>](https://reviewable.io/reviews/kkrt-labs/kakarot/1199)
<!-- Reviewable:end -->