fix: use immutable in L1KakarotMessaging (#1242)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days.
Did you spend 0.5 days on this PR or rather 2 days?  -->

Time spent on this PR:

## Pull request type

<!-- Please try to limit your pull request to one type,
submit multiple pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?
`consumeMessageFromL2` in `L1KakarotMessaging` was not used and
commented.
There is a storage collision when calling `L1KakarotMessaging` in
delegatecall as `_starknetMessaging` and `_kakarotAddress` are stored in
slot 0 and 1.

<!-- Please describe the current behavior that you are modifying,
or link to a relevant issue. -->

Resolves #1228

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->
- In `L1KakarotMessaging` contract 
    - `_starknetMessaging` and `_kakarotAddress` are passed as immutable
This contract should not have storage if possible to avoid storage
collision. If necessary implement
[ERC7201](https://eips.ethereum.org/EIPS/eip-7201) or other solution.
- adds a `self` immutable to add a check in `consumeMessageFromL2` to
ensure it is called in `delegatecall`
- forge fmt

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg"
height="34" align="absmiddle"
alt="Reviewable"/>](https://reviewable.io/reviews/kkrt-labs/kakarot/1242)
<!-- Reviewable:end -->